requirements.txt
python-telegram-bot==20.7
ccxt==4.0.96
aiohttp==3.9.1
google-generativeai==0.5.0
groq==0.4.2
requests==2.31.0
schedule==1.2.0
pytz==2024.1
python-dotenv==1.0.0
asyncio==3.4.3
